requirements.txt
python_dotenv
langchain
langchain-community
langchain_text_splitters
langchain_openai
langchain_chroma